Add support for MIDI in devices
This introduces a new Lua function "lmc_set_midi_handler" which can be used to map midi input devices. The devices are managed by a midideviceservice, similar to keyboards and game controllers.

The handler function in Lua will receive the midi input data (which is status, data1 and data2).

Currently no support for SysEx.